Refactor client auth handling in oauth and passwordless
